+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- new `numeric.solver` namespace with a shim over Apache Commons Math's ODE solvers
+- dependency on `org.apache.commons/commons-math3`
 
 ## [0.0.15] — 2025-01-16
 ### Fixed